UIE-204 Narrow Ajax Usage pt19 - narrowed Ajax() usage within more of src/workflows-app area to call Ajax().SubAreaX directly. - updated rest of the shared components and related tests. - improved mock types where possible. - light Typescript conversions to assist in safer code udpates. - additional Typescript improvements around component props.
