+++ src/libs/ajax/workflows-app/WorkflowScript.ts
@@ -1,8 +1,10 @@
 import { fetchOk } from 'src/libs/ajax/fetch/fetch-core';
 
-export const WorkflowScript = (signal) => ({
+export const WorkflowScript = (signal?: AbortSignal) => ({
   get: async (workflowUrl) => {
     const res = await fetchOk(workflowUrl, { signal, method: 'GET' });
     return res.text();
   },
 });
+
+export type WorkflowScriptAjaxContract = ReturnType<typeof WorkflowScript>;
