+++ src/libs/state.ts
@@ -368,12 +368,24 @@ export const AppProxyUrlStatus = Object.freeze({
   Error: 'Error',
 });
 
+export interface WorkflowsProxyUrlState {
+  status: keyof typeof AppProxyUrlStatus;
+  state: string;
+}
+
+export interface WorkflowsAppStoreState {
+  workspaceId?: string;
+  wdsProxyUrlState: WorkflowsProxyUrlState;
+  cbasProxyUrlState: WorkflowsProxyUrlState;
+  cromwellProxyUrlState: WorkflowsProxyUrlState;
+}
+
 /*
  * Stores the proxy urls for WDS and Azure Workflows apps for a workspace.
  * Status can be one of None, Ready and Error. The proxy url will be in 'state' field when 'status' is Ready.
  * When 'state' is Error the 'state' field will contain the error that was returned from Leo (if any).
  */
-export const workflowsAppStore = atom({
+export const workflowsAppStore = atom<WorkflowsAppStoreState>({
   workspaceId: undefined,
   wdsProxyUrlState: { status: AppProxyUrlStatus.None, state: '' },
   cbasProxyUrlState: { status: AppProxyUrlStatus.None, state: '' },
