+++ src/workflows-app/components/WorkflowsAppNavPanel.test.ts
@@ -2,8 +2,7 @@ import { act, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { h } from 'react-hyperscript-helpers';
 import { AnalysesData } from 'src/analysis/Analyses';
-import { Ajax } from 'src/libs/ajax';
-import { asMockedFn, renderWithAppContexts as render } from 'src/testing/test-utils';
+import { renderWithAppContexts as render } from 'src/testing/test-utils';
 import { WorkflowsAppNavPanel } from 'src/workflows-app/components/WorkflowsAppNavPanel';
 import { mockAzureWorkspace } from 'src/workflows-app/utils/mock-responses';
 
@@ -23,30 +22,21 @@ const defaultAnalysesDataWithAppsRefreshed: AnalysesData = {
   lastRefresh: new Date(),
 };
 
-jest.mock('src/libs/ajax');
+type UseMetricsExports = typeof import('src/libs/ajax/metrics/useMetrics');
+jest.mock(
+  'src/libs/ajax/metrics/useMetrics',
+  (): UseMetricsExports => ({
+    ...jest.requireActual<UseMetricsExports>('src/libs/ajax/metrics/useMetrics'),
+    useMetricsEvent: jest.fn(() => ({ captureEvent: jest.fn() })),
+  })
+);
 
 jest.mock('src/libs/nav', () => ({
   ...jest.requireActual('src/libs/nav'),
   useQueryParameter: jest.requireActual('react').useState,
 }));
 
-const watchCaptureEvent = jest.fn();
-type AjaxContract = ReturnType<typeof Ajax>;
-type AjaxMetricsContract = AjaxContract['Metrics'];
-const mockMetrics: Partial<AjaxMetricsContract> = {
-  captureEvent: (event, details) => watchCaptureEvent(event, details),
-};
-
-const defaultAjaxImpl: Partial<AjaxContract> = {
-  Metrics: mockMetrics as AjaxMetricsContract,
-};
-
 describe('Workflows App Navigation Panel', () => {
-  beforeEach(() => {
-    // Arrange
-    asMockedFn(Ajax).mockReturnValue(defaultAjaxImpl as AjaxContract);
-  });
-
   it('renders headers', async () => {
     const user = userEvent.setup();
 
@@ -62,7 +52,6 @@ describe('Workflows App Navigation Panel', () => {
         workspace: mockAzureWorkspace,
         analysesData: defaultAnalysesData,
         setLoading: jest.fn(),
-        signal: jest.fn(),
       })
     );
 
@@ -101,7 +90,6 @@ describe('Workflows App Navigation Panel', () => {
         workspace: mockAzureWorkspace,
         analysesData: defaultAnalysesData,
         setLoading: jest.fn(),
-        signal: jest.fn(),
       })
     );
 
@@ -133,7 +121,6 @@ describe('Workflows App Navigation Panel', () => {
           },
           analysesData: defaultAnalysesData,
           setLoading: jest.fn(),
-          signal: jest.fn(),
         })
       )
     );
@@ -166,7 +153,6 @@ describe('Workflows App Navigation Panel', () => {
         workspace: mockAzureWorkspace,
         analysesData: defaultAnalysesData,
         setLoading: jest.fn(),
-        signal: jest.fn(),
       })
     );
 
@@ -186,7 +172,6 @@ describe('Workflows App Navigation Panel', () => {
         workspace: mockAzureWorkspace,
         analysesData: defaultAnalysesDataWithAppsRefreshed,
         setLoading: jest.fn(),
-        signal: jest.fn(),
       })
     );
 
@@ -206,7 +191,6 @@ describe('Workflows App Navigation Panel', () => {
           workspace: mockAzureWorkspace,
           analysesData: defaultAnalysesDataWithAppsRefreshed,
           setLoading: jest.fn(),
-          signal: jest.fn(),
         })
       );
     });
@@ -278,7 +262,6 @@ describe('Workflows App Navigation Panel', () => {
           workspace: mockAzureWorkspace,
           analysesData: analysesDataWithAppsInError,
           setLoading: jest.fn(),
-          signal: jest.fn(),
         })
       );
     });
